login and sign-up authentication for user backend

+++ backend/src/services/posts.js
@@ -1,6 +1,6 @@
 import { Post } from '../db/models/post.js'
- export async function createPost({ title, author, contents, tags }) { 
-    const post = new Post({ title, author, contents, tags }) 
+ export async function createPost( userID, {title, contents, tags}) { 
+    const post = new Post({ title, author: userID, contents, tags }) 
     return await post.save() 
 }
 
@@ -26,11 +26,11 @@ export async function getPostById(postId) {
     return await Post.findById(postId) 
 }
 
-export async function updatePost(postId, { title, author, contents, tags }) { 
-    return await Post.findOneAndUpdate( { _id: postId }, { $set: { title, author, contents, tags } },
+export async function updatePost(userID, postId, { title, contents, tags }) { 
+    return await Post.findOneAndUpdate( { _id: postId, author:userID }, { $set: { title, contents, tags } },
         { new: true }, ) 
     }
 
-export async function deletePost(postId) { 
-    return await Post.deleteOne({ _id: postId }) 
+export async function deletePost(userID, postId) { 
+    return await Post.deleteOne({ _id: postId, author: userID }) 
 }
